requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
gspread==5.12.0
google-auth==2.23.4
python-dateutil==2.8.2
python-dotenv==1.0.0
